refactor: group stories
